+++ src/app/dim-api/reducer.ts
@@ -11,12 +11,14 @@ import {
   TagValue,
 } from '@destinyitemmanager/dim-api-types';
 import { DestinyAccount } from 'app/accounts/destiny-account';
+import { recentSearchComparator } from 'app/search/autocomplete';
 import { canonicalizeQuery, parseQuery } from 'app/search/query-parser';
 import { searchConfigSelector } from 'app/search/search-config';
 import { validateQuery } from 'app/search/search-utils';
 import { RootState } from 'app/store/types';
 import { emptyArray } from 'app/utils/empty';
 import { errorLog, infoLog, timer } from 'app/utils/log';
+import { count } from 'app/utils/util';
 import { clearWishLists } from 'app/wishlists/actions';
 import produce, { Draft } from 'immer';
 import _ from 'lodash';
@@ -32,6 +34,9 @@ import { initialSettingsState, Settings } from '../settings/initial-settings';
 import { DeleteLoadoutUpdateWithRollback, ProfileUpdateWithRollback } from './api-types';
 import * as actions from './basic-actions';
 import { makeProfileKey, makeProfileKeyFromAccount } from './selectors';
+
+// After you've got a search history of more than this many items, we start deleting the older ones
+const MAX_SEARCH_HISTORY = 300;
 
 export interface DimApiState {
   globalSettings: GlobalSettings;
@@ -197,7 +202,8 @@ export const dimApi = (
 
     case getType(actions.profileLoaded): {
       const { profileResponse, account } = action.payload;
-      return {
+      // TODO: clean out invalid/simple searches on first load?
+      const newState = {
         ...state,
         profileLoaded: true,
         profileLoadedError: undefined,
@@ -227,6 +233,13 @@ export const dimApi = (
             }
           : state.searches,
       };
+
+      // If this is the first load, cleanup searches
+      if (account && !state.profileLoaded) {
+        return produce(newState, (state) => cleanupInvalidSearches(state, account));
+      }
+
+      return newState;
     }
 
     case getType(actions.profileLoadError): {
@@ -1012,7 +1025,6 @@ function searchUsed(draft: Draft<DimApiState>, account: DestinyAccount, query: s
     }
     if (ast.op === 'noop' || (ast.op === 'filter' && ast.type === 'keyword')) {
       // don't save "trivial" single-keyword filters
-      // TODO: somehow also reject invalid searches (that don't match real keywords)
       return;
     }
     query = canonicalizeQuery(ast);
@@ -1044,11 +1056,20 @@ function searchUsed(draft: Draft<DimApiState>, account: DestinyAccount, query: s
     });
   }
 
-  // TODO: purge invalid searches
-  // TODO: this is where we would cap the search history!
-  // while (searches.length > MAX_SEARCH_HISTORY) {
-  //   remove bottom-sorted search
-  // }
+  if (searches.length > MAX_SEARCH_HISTORY) {
+    const sortedSearches = [...searches].sort(recentSearchComparator);
+
+    const numBuiltinSearches = count(sortedSearches, (s) => s.usageCount <= 0);
+
+    // remove bottom-sorted search until we get to the limit
+    while (sortedSearches.length > MAX_SEARCH_HISTORY - numBuiltinSearches) {
+      const lastSearch = sortedSearches.pop()!;
+      // Never try to delete the built-in searches or saved searches
+      if (!lastSearch.saved && lastSearch.usageCount > 0) {
+        deleteSearch(draft, destinyVersion, lastSearch.query);
+      }
+    }
+  }
 
   draft.updateQueue.push(updateAction);
 }
@@ -1122,6 +1143,37 @@ function deleteSearch(draft: Draft<DimApiState>, destinyVersion: DestinyVersion,
   draft.searches[destinyVersion] = draft.searches[destinyVersion].filter((s) => s.query !== query);
 
   draft.updateQueue.push(updateAction);
+}
+
+function cleanupInvalidSearches(draft: Draft<DimApiState>, account: DestinyAccount) {
+  // Filter out saved and builtin searches
+  const searches = draft.searches[account.destinyVersion].filter(
+    (s) => !s.saved && s.usageCount > 0
+  );
+
+  if (!searches.length) {
+    return;
+  }
+
+  const searchConfigs = searchConfigSelector(stubSearchRootState(account));
+  for (const search of draft.searches[account.destinyVersion]) {
+    if (search.saved || search.usageCount <= 0) {
+      continue;
+    }
+
+    try {
+      const ast = parseQuery(search.query);
+      if (
+        !validateQuery(ast, searchConfigs) ||
+        ast.op === 'noop' ||
+        (ast.op === 'filter' && ast.type === 'keyword')
+      ) {
+        deleteSearch(draft, account.destinyVersion, search.query);
+      }
+    } catch (e) {
+      deleteSearch(draft, account.destinyVersion, search.query);
+    }
+  }
 }
 
 function reverseEffects(draft: Draft<DimApiState>, update: ProfileUpdateWithRollback) {